feat: implement VcChip to InStock & CountInCart
