+++ client-app/shared/cart/components/cart-line-items.vue
@@ -31,7 +31,7 @@
           @update:model-value="$emit('change:itemQuantity', { itemId: item.id, quantity: $event })"
         />
 
-        <VcInStock
+        <InStock
           :is-in-stock="item.availabilityData?.isInStock"
           :is-available="!item.deleted"
           :quantity="item.availabilityData?.availableQuantity"
@@ -64,6 +64,7 @@ import { computed } from "vue";
 import { useCartValidationErrorTranslator } from "@/core/composables";
 import { ProductType } from "@/core/enums";
 import { prepareLineItems } from "@/core/utilities";
+import { InStock } from "@/shared/catalog";
 import type { LineItemType, ValidationErrorType } from "@/core/api/graphql/types";
 
 interface IProps {
