+++ client-app/shared/catalog/components/count-in-cart.vue
@@ -1,11 +1,8 @@
 <template>
-  <div
-    v-if="countInCart > 0"
-    class="whitespace-nowrap rounded-full bg-[color:var(--color-add-to-cart-in-cart-bg)] px-[0.677rem] py-0.5 text-[13px] leading-5 text-[color:var(--color-add-to-cart-in-cart)] lg:px-[0.53rem] lg:py-px lg:text-[11px]"
-  >
-    <span class="inline-block text-center font-bold">{{ countInCart }}</span>
+  <VcChip v-if="countInCart > 0" size="sm" variant="outline-dark" color="neutral" rounded>
+    {{ countInCart }}
     {{ $t("shared.cart.add_to_cart.errors.in_cart") }}
-  </div>
+  </VcChip>
 </template>
 
 <script setup lang="ts">
