+++ client-app/shared/catalog/components/index.ts
@@ -1,9 +1,11 @@
 export { default as CarouselProductCard } from "./carousel-product-card.vue";
 export { default as Category } from "./category.vue";
 export { default as CategorySelector } from "./category-selector.vue";
+export { default as CountInCart } from "./count-in-cart.vue";
 export { default as DiscountBadge } from "./discount-badge.vue";
 export { default as DisplayProducts } from "./display-products.vue";
 export { default as FacetFilter } from "./facet-filter.vue";
+export { default as InStock } from "./in-stock.vue";
 export { default as ProductCardCompare } from "./product-card-compare.vue";
 export { default as ProductCardGrid } from "./product-card-grid.vue";
 export { default as ProductCardList } from "./product-card-list.vue";
