+++ client-app/shared/catalog/components/product-card-list.vue
@@ -137,13 +137,13 @@
         <slot name="cart-handler"></slot>
 
         <div class="flex items-center gap-1 lg:mt-0.5">
-          <VcInStock
+          <InStock
             :is-in-stock="product.availabilityData?.isInStock"
             :is-digital="isDigital"
             :quantity="product.availabilityData?.availableQuantity"
           />
 
-          <VcCountInCart :product-id="product.id" />
+          <CountInCart :product-id="product.id" />
         </div>
       </template>
     </div>
@@ -156,7 +156,9 @@ import { ProductType, PropertyType } from "@/core/enums";
 import { getLinkTarget, getProductRoute, getPropertiesGroupedByName } from "@/core/utilities";
 import { AddToCompareCatalog } from "@/shared/compare";
 import { AddToList } from "@/shared/wishlists";
+import CountInCart from "./count-in-cart.vue";
 import DiscountBadge from "./discount-badge.vue";
+import InStock from "./in-stock.vue";
 import Vendor from "./vendor.vue";
 import type { Product } from "@/core/api/graphql/types";
