+++ client-app/ui-kit/components/atoms/index.ts
@@ -9,7 +9,6 @@ export { default as VcCardSkeleton } from "./card-skeleton/vc-card-skeleton.vue"
 export { default as VcCardWidget } from "./card-widget/vc-card-widget.vue";
 export { default as VcCheckbox } from "./checkbox/vc-checkbox.vue";
 export { default as VcContainer } from "./container/vc-container.vue";
-export { default as VcCountInCart } from "./count-in-cart/vc-count-in-cart.vue";
 export { default as VcDateSelector } from "./date-selector/vc-date-selector.vue";
 export { default as VcEmptyView } from "./empty-view/vc-empty-view.vue";
 export { default as VcExpansionPanels } from "./expansion-panels/vc-expansion-panels.vue";
@@ -18,7 +17,6 @@ export { default as VcFilterCardSkeleton } from "./filter-card-skeleton/vc-filte
 export { default as VcHexagonIcon } from "./hexagon-icon/vc-hexagon-icon.vue";
 export { default as VcIcon } from "./icon/vc-icon.vue";
 export { default as VcImage } from "./image/vc-image.vue";
-export { default as VcInStock } from "./in-stock/vc-in-stock.vue";
 export { default as VcInfinityScrollLoader } from "./infinity-scroll-loader/vc-infinity-scroll-loader.vue";
 export { default as VcInputDetails } from "./input-details/vc-input-details.vue";
 export { default as VcLabel } from "./label/vc-label.vue";
